+++ python/cudf/cudf/core/index.py
@@ -145,6 +145,12 @@ class Index(Frame, Serializable):
         return idx_typ(index, name=name)
 
     @property
+    def ndim(self):
+        """Dimension of the data. Apart from MultiIndex ndim is always 1.
+        """
+        return 1
+
+    @property
     def names(self):
         """
         Returns a tuple containing the name of the Index.